+++ Classes/Domain/Model/Tool.php
@@ -10,10 +10,7 @@ declare(strict_types=1);
 
 namespace Slavlee\FoodRecipes\Domain\Model;
 
-use GeorgRinger\News\Domain\Model\News;
-use TYPO3\CMS\Extbase\Annotation\ORM\Lazy;
 use TYPO3\CMS\Extbase\DomainObject\AbstractEntity;
-use TYPO3\CMS\Extbase\Persistence\ObjectStorage;
 
 class Tool extends AbstractEntity
 {